rename common to cli.common (not used outside cli)
